requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
